Rename batch item
