+++ src/content/index.ts
@@ -1,4 +1,4 @@
-export * from "./BatchItem.js";
+export * from "./BatchRequestStep";
 export * from "./BatchRequestContent.js";
 export * from "./BatchRequestContentCollection.js";
 export * from "./BatchResponseContent.js";
